Changed some assert(sizeof(x)) to static_assert's. Fixed problem with test suite.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@6333 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/StringUtils.h
@@ -35,11 +35,11 @@ uint64 stringToUInt64(const std::string& s); // throws StringUtilsExcep
 
 unsigned int hexCharToUInt(char c);
 unsigned int hexStringToUInt(const std::string& s);
-unsigned long long hexStringTo64UInt(const std::string& s);
+uint64 hexStringTo64UInt(const std::string& s);
 
 //====================== Number to String conversion ======================
 //const std::string toHexString(unsigned int i);//32 bit integers
-const std::string toHexString(unsigned long long i);//for 64 bit integers
+const std::string toHexString(uint64 i);//for 64 bit integers
 
 const std::string int32ToString(int32 i);
 const std::string int64ToString(int64 i);
